// BuildingAssociation/client/src/models/ConsumptionType.ts
//     This code was generated by a Reinforced.Typings tool. 
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import * as CalculationType from './CalculationType';

export interface ConsumptionType
{
	id?: number;
	name: string;
	calculationType: CalculationType.CalculationType;
	date: string;
	mansionId?: number;
	mansionName: string;
}